Introduce a mechanism for services to be definable outside of the core binary and to depend on them asynchronously. Use it for the Cid service that will be defined in amp-analytics.
