+++ build-system/tasks/presubmit-checks.js
@@ -48,6 +48,7 @@ var forbiddenTerms = {
     message: requiresReviewPrivacy,
     whitelist: [
       'src/cid.js',
+      'src/service/cid-impl.js',
       'test/functional/test-cid.js',
     ],
   },
@@ -55,7 +56,7 @@ var forbiddenTerms = {
     message: requiresReviewPrivacy,
     whitelist: [
       'src/cookies.js',
-      'src/cid.js',
+      'src/service/cid-impl.js',
       'test/functional/test-cid.js',
       'test/functional/test-cookies.js',
       'test/functional/test-experiments.js',
@@ -64,7 +65,7 @@ var forbiddenTerms = {
   'getCookie\\W': {
     message: requiresReviewPrivacy,
     whitelist: [
-      'src/cid.js',
+      'src/service/cid-impl.js',
       'src/cookies.js',
       'src/experiments.js',
       'test/functional/test-cookies.js',
@@ -85,7 +86,7 @@ var forbiddenTerms = {
     message: requiresReviewPrivacy,
     whitelist: [
       'test/_init_tests.js',
-      'src/cid.js',
+      'src/service/cid-impl.js',
       'test/functional/test-cid.js',
     ],
   },
